Move gltf stuff in pbr to its own file
